+++ src/main/java/me/dominiksta/jfiredump/DBExporter.java
@@ -1,8 +1,9 @@
 package me.dominiksta.jfiredump;
 
 import java.io.BufferedWriter;
-import java.io.FileWriter;
+import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.OutputStreamWriter;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
@@ -16,8 +17,17 @@ public abstract class DBExporter {
     /** A connection to the database to export from */
     protected DBConnection con;
 
+    /** Newline character to use in export */
+    protected String nl = "\n";
+
     public DBExporter(DBConnection con) {
         this.con = con;
+
+        // https://github.com/FirebirdSQL/jaybird/wiki/Character-encodings
+        String lowerEnc = this.con.getEncoding().toLowerCase();
+        if (lowerEnc.startsWith("win") || lowerEnc.startsWith("dos")) {
+            this.nl = "\r\n";
+        }
     }
 
     /**
@@ -33,7 +43,11 @@ public abstract class DBExporter {
     /** Helper to return an open BufferedWriter for `fileName` */
     protected BufferedWriter writerForPath(String fileName) {
         try {
-            return new BufferedWriter(new FileWriter(fileName));
+            return new BufferedWriter(
+                new OutputStreamWriter(
+                    new FileOutputStream(fileName), "UTF-8"
+                )
+            );
         } catch(IOException e) {
             App.logger.severe("Could not open file with path " + fileName);
             e.printStackTrace();
@@ -48,5 +62,13 @@ public abstract class DBExporter {
             new SimpleDateFormat("yyyy-MM-dd_HH-mm-ss").format(new Date()),
             tableName
         );
+    }
+
+    public String getNewline() {
+        return nl;
+    }
+
+    public void setNewline(String nl) {
+        this.nl = nl;
     }
 }